refactor dashboards into ui package
