--- apps/main-site/src/components/use-domain-status.tsx
@@ -1,18 +1,0 @@
-import { trpc } from '@answeroverflow/ui';
-
-export function useDomainStatus({ domain }: { domain?: string }) {
-	const { data, isLoading, isFetching } =
-		trpc.servers.verifyCustomDomain.useQuery(domain ?? '', {
-			refetchOnMount: true,
-			refetchInterval: 5000,
-			keepPreviousData: true,
-			enabled: domain !== undefined,
-		});
-
-	return {
-		status: data?.status,
-		domainJson: data?.domainJson,
-		loading: isLoading,
-		fetching: isFetching,
-	};
-}
