+++ packages/ui/src/components/dashboard/domains.tsx
@@ -1,14 +1,37 @@
-import { useState } from 'react';
-import { useDomainStatus } from './use-domain-status';
-import { LuAlertCircle, LuXCircle } from 'react-icons/lu';
 import {
+	Card,
 	Title,
+	Subtitle,
 	TabGroup,
-	Tab,
 	TabList,
-	TabPanel,
+	Tab,
 	TabPanels,
+	TabPanel,
 } from '@tremor/react';
+import type { ServerPublic } from '@answeroverflow/api';
+import { useState } from 'react';
+import { LuAlertCircle, LuXCircle, LuCheckCircle2 } from 'react-icons/lu';
+import { toast } from 'react-toastify';
+import { trpc } from '~ui/utils/trpc';
+import { Button, Input, LoadingSpinner } from '../primitives';
+import { useTierAccess } from '../primitives/tier-access-only';
+
+export function useDomainStatus({ domain }: { domain?: string }) {
+	const { data, isLoading, isFetching } =
+		trpc.servers.verifyCustomDomain.useQuery(domain ?? '', {
+			refetchOnMount: true,
+			refetchInterval: 5000,
+			keepPreviousData: true,
+			enabled: domain !== undefined,
+		});
+
+	return {
+		status: data?.status,
+		domainJson: data?.domainJson,
+		loading: isLoading,
+		fetching: isFetching,
+	};
+}
 
 export const getSubdomain = (name: string, apexName: string) => {
 	if (name === apexName) return null;
@@ -21,7 +44,98 @@ const InlineSnippet = ({ children }: { children: string }) => {
 		</span>
 	);
 };
-export default function DomainConfiguration({ domain }: { domain: string }) {
+
+export function ConfigureDomainCard(props: {
+	server: Pick<ServerPublic, 'id' | 'customDomain'>;
+}) {
+	const { enabled } = useTierAccess();
+	const { customDomain: currentDomain, id } = props.server;
+	const util = trpc.useContext();
+	const mutation = trpc.servers.setCustomDomain.useMutation({
+		onSuccess: () => {
+			toast.success('Custom domain updated!');
+			void util.servers.fetchDashboardById.invalidate(id);
+		},
+		onError: (err) => {
+			toast.error(err.message);
+		},
+	});
+	const { fetching } = useDomainStatus({ domain: currentDomain ?? undefined });
+
+	return (
+		<Card>
+			<form
+				onSubmit={(e) => {
+					e.preventDefault();
+					// @ts-ignore
+					// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+					const newDomain = e.target[1].value as string;
+					if (
+						currentDomain &&
+						newDomain !== currentDomain &&
+						!confirm('Are you sure you want to change your custom domain?')
+					) {
+						return;
+					}
+
+					mutation.mutate({
+						customDomain: newDomain,
+						serverId: id,
+					});
+				}}
+			>
+				<div className="relative flex flex-col space-y-4 p-5">
+					<div className="flex items-center justify-between">
+						<div>
+							<Title>Custom domain</Title>
+							<Subtitle>The custom domain for your site.</Subtitle>
+						</div>
+						<Button
+							onClick={() => {
+								void util.servers.verifyCustomDomain.invalidate();
+							}}
+							type="button"
+							disabled={!enabled || fetching}
+							variant={'outline'}
+							className="relative"
+						>
+							{fetching && (
+								<>
+									<div className="absolute right-16 z-10 flex h-full items-center">
+										<LoadingSpinner />
+									</div>
+									<div className="w-6" />
+								</>
+							)}
+							Refresh
+						</Button>
+					</div>
+					<div className="relative flex w-full max-w-md">
+						<Input
+							name="customDomain"
+							type="text"
+							key={currentDomain ?? props.server.id}
+							defaultValue={currentDomain || ''}
+							placeholder="yourdomain.com"
+							maxLength={64}
+							disabled={!enabled}
+							pattern={'[a-z0-9]+([\\-\\.]{1}[a-z0-9]+)*\\.[a-z]{2,5}$'}
+						/>
+						{currentDomain && <DomainStatus domain={currentDomain} />}
+					</div>
+				</div>
+				{currentDomain && <DomainConfiguration domain={currentDomain} />}
+				<div className="flex flex-col items-center justify-center space-y-2 rounded-b-lg border-t border-stone-200 bg-stone-50 p-3 dark:bg-stone-900 sm:flex-row sm:justify-between sm:space-y-0 sm:px-6">
+					<p className="text-sm text-stone-500 dark:text-stone-400">
+						Please enter a valid domain.
+					</p>
+				</div>
+			</form>
+		</Card>
+	);
+}
+
+export function DomainConfiguration({ domain }: { domain: string }) {
 	const [recordType, setRecordType] = useState<'A' | 'CNAME'>('A');
 	const { status, domainJson } = useDomainStatus({ domain });
 
@@ -160,6 +274,24 @@ export default function DomainConfiguration({ domain }: { domain: string }) {
 						</p>
 					</div>
 				</>
+			)}
+		</div>
+	);
+}
+
+export function DomainStatus({ domain }: { domain: string }) {
+	const { status, loading } = useDomainStatus({ domain });
+
+	return (
+		<div className="absolute right-3 z-10 flex h-full items-center">
+			{loading ? (
+				<LoadingSpinner />
+			) : status === 'Valid Configuration' ? (
+				<LuCheckCircle2 fill="#2563EB" stroke="white" />
+			) : status === 'Pending Verification' ? (
+				<LuAlertCircle fill="#FBBF24" stroke="white" />
+			) : (
+				<LuXCircle fill="#DC2626" stroke="white" />
 			)}
 		</div>
 	);
