+++ packages/ui/src/components/dashboard/index.tsx
@@ -1,0 +1,1 @@
+export * from './server-dashboard';
