+++ packages/ui/src/components/dashboard/server-dashboard.tsx
@@ -1,0 +1,108 @@
+import {
+	TabGroup,
+	TabList,
+	Tab,
+	TabPanels,
+	TabPanel,
+	Grid,
+	Card,
+	Col,
+	Title,
+} from '@tremor/react';
+import { AOHead, GetStarted, AOLink, trpc } from '@answeroverflow/ui';
+import { ConfigureDomainCard } from './domains';
+import { GoLinkExternal } from 'react-icons/go';
+import { DashboardNavbar } from './dashboard-navbar';
+import { CurrentPlanCard, LineChartCard, PageViewsCard } from './cards';
+import { TierAccessOnly } from '../primitives/tier-access-only';
+
+export function ServerDashboard(props: { serverId: string }) {
+	const { data, status } = trpc.servers.fetchDashboardById.useQuery(
+		props.serverId,
+	);
+
+	const Body = () => {
+		switch (status) {
+			case 'loading':
+				return (
+					<div className="flex h-[50vh] items-center justify-center">
+						<div className="h-32 w-32 animate-spin rounded-full border-b-4 border-ao-blue" />
+					</div>
+				);
+			case 'error':
+				return (
+					<div className="flex h-screen flex-col items-center justify-center">
+						<div className="text-4xl font-bold">No servers with bot found</div>
+						<div className="text-2xl">
+							Add the bot to a server to get started
+						</div>
+						<GetStarted className="mt-6" location="Pricing" />
+					</div>
+				);
+			case 'success':
+				return (
+					<TabGroup className="px-2 md:px-8">
+						<TabList className="mx-auto max-w-7xl">
+							<Tab>Overview</Tab>
+						</TabList>
+						<TabPanels className="mx-auto max-w-7xl">
+							<TabPanel>
+								<Grid numItemsLg={6} className="mt-6 gap-6">
+									{/* Main section */}
+									<Col numColSpanLg={4}>
+										<LineChartCard serverId={data.id} />
+									</Col>
+
+									{/* KPI sidebar */}
+									<Col
+										numColSpanLg={2}
+										className="flex flex-col justify-between gap-4"
+									>
+										<Card>
+											<AOLink
+												href={
+													data.customDomain
+														? `https://${data.customDomain}`
+														: `/c/${data.id}`
+												}
+												className="flex items-center gap-2"
+											>
+												<Title>View Community</Title>
+												<GoLinkExternal className="mt-1 h-5 w-5" />
+											</AOLink>
+										</Card>
+										<PageViewsCard serverId={data.id} />
+
+										<CurrentPlanCard
+											server={data}
+											stripeUrl={data.stripeCheckoutUrl}
+											dateCancelationTakesEffect={data.cancelAt}
+											dateSubscriptionRenews={data.currentPeriodEnd}
+											dateTrialEnds={data.trialEnd}
+										/>
+									</Col>
+								</Grid>
+
+								<div className="mt-6">
+									<TierAccessOnly
+										enabledFor={['PRO', 'OPEN_SOURCE']}
+										currentPlan={data.plan}
+									>
+										<ConfigureDomainCard server={data} />
+									</TierAccessOnly>
+								</div>
+							</TabPanel>
+						</TabPanels>
+					</TabGroup>
+				);
+		}
+	};
+
+	return (
+		<>
+			<AOHead title="Dashboard" path={'/dashboard'} />
+			<DashboardNavbar />
+			<Body />
+		</>
+	);
+}
