+++ packages/ui/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './primitives';
 export * from './pages';
+export * from './dashboard';
